update_footer
